Convert to float32 before plotting pianorolls

+++ pianoroll_utils.py
@@ -42,7 +42,7 @@ def plot_pianoroll(ax, pianoroll, min_pitch=0, max_pitch=127, beat_resolution=No
     assert pianoroll.shape[0] == max_pitch - min_pitch + 1
     num_ticks = pianoroll.shape[1]
     
-    ax.imshow(pianoroll, cmap=cmap, aspect='auto', 
+    ax.imshow(pianoroll.astype('float32'), cmap=cmap, aspect='auto', 
               vmin=0, vmax=1, origin='lower', interpolation='none')
     ax.set_ylabel('pitch')
     lowest_octave = ((min_pitch - 1) // 12 + 1) - 2
